requirements.txt
numpy==2.0.1
pandas==2.2.2
streamlit==1.37.1
streamlit-extras==0.4.7
streamlit-pdf-reader==0.0.5
plotly==5.23.0
python-dotenv==1.0.1
streamlit_authenticator==0.3.3
google-api-python-client==2.144.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
openpyxl==3.1.2
pymongo[srv]==4.6.1
